# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ServicesOnLineAPI",
    "version": "v1"
  },
  "paths": {
...
    "/api/person": {
      "get": {
        "tags": [
          "Person"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Person"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/person/{id}": {
      "get": {
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/person/{pesonId}": {
      "put": {
        "tags": [
          "Person"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pesonId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pesonId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "PersonInput": {
      "type": "object",
      "properties": {
        "personName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "roomId": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}